Refactor WPF right pane
