+++ HomeInventory/HomeInventory.Desktop.Wpf/Views/RightPaneView.xaml
@@ -1,78 +1,22 @@
-﻿<UserControl x:Class="HomeInventory.Desktop.Wpf.Views.RightPaneView"
+<UserControl x:Class="HomeInventory.Desktop.Wpf.Views.RightPaneView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:xctk="http://schemas.xceed.com/wpf/xaml/toolkit"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:vm="clr-namespace:HomeInventory.Desktop.Wpf.ViewModels"
-             xmlns:local="clr-namespace:HomeInventory.Desktop.Wpf.Views"
-             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800"
-             >
-    <UserControl.Resources>
-        <BooleanToVisibilityConverter x:Key="BoolToVisibilityConverter"/>
-        <Style x:Key="NameTemplate" TargetType="{x:Type TextBox}">
-            <Setter Property="MaxLength" Value="20" />
-        </Style>
-        <Style x:Key="DescPlaceNoteTemplate" TargetType="{x:Type TextBox}">
-            <Setter Property="MaxLength" Value="100" />
-        </Style>
-    </UserControl.Resources>
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             mc:Ignorable="d"
+             d:DesignHeight="450" d:DesignWidth="800">
     <Grid>
         <Grid.RowDefinitions>
-            <RowDefinition Height="30" />
-            <RowDefinition Height="*" />
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="*"/>
         </Grid.RowDefinitions>
-        <Grid Grid.Row="0">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="*" />
-                <ColumnDefinition Width="Auto" />
-                <ColumnDefinition Width="Auto" />
-                <ColumnDefinition Width="Auto" />
-            </Grid.ColumnDefinitions>
-            <Button Command="{Binding AddItemCommand}" Margin="0,0,8,0" Grid.Column="1">
-                <materialDesign:PackIcon Kind="Add" Width="20" Height="20"/>
-            </Button>
-            <Button Command="{Binding DeleteItemCommand}" Margin="0,0,8,0" Grid.Column="2">
-                <materialDesign:PackIcon Kind="Delete" Width="20" Height="20"/>
-            </Button>
-            <Button Command="{Binding MoveToLocationCommand}" Margin="0,0,8,0" Grid.Column="3">
-                <materialDesign:PackIcon Kind="SwapHorizontal" Width="20" Height="20"/>
-            </Button>
-        </Grid>
-        <DataGrid ItemsSource="{Binding Items}" AutoGenerateColumns="False" SelectedItem="{Binding SelectedItem}" CanUserAddRows="False" x:Name="ItemsGrid" Loaded="DataGrid_Loaded" IsReadOnly="{Binding IsBusy}"  Grid.Row="1">
-            
-            <DataGrid.Columns>
-                <DataGridCheckBoxColumn Binding="{Binding IsSelected, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" IsThreeState = "False"/>
 
-                <DataGridTextColumn EditingElementStyle="{StaticResource NameTemplate}" Header="Name" Binding="{Binding Name, UpdateSourceTrigger=LostFocus}" />
-                <DataGridTemplateColumn Header="Quantity">
-                    <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate>
-                            <TextBlock Text="{Binding Quantity}" />
-                        </DataTemplate>
-                    </DataGridTemplateColumn.CellTemplate>
-                    <DataGridTemplateColumn.CellEditingTemplate>
-                        <DataTemplate>
-                            <xctk:IntegerUpDown Value="{Binding Quantity, Mode=TwoWay, UpdateSourceTrigger=LostFocus}" Minimum="0" FormatString="N0" />
-                        </DataTemplate>
-                    </DataGridTemplateColumn.CellEditingTemplate>
-                </DataGridTemplateColumn>
-                <DataGridTextColumn EditingElementStyle="{StaticResource DescPlaceNoteTemplate}" Header="Placement Note" Binding="{Binding PlacementNote, UpdateSourceTrigger=LostFocus}" />
-                <DataGridTextColumn EditingElementStyle="{StaticResource DescPlaceNoteTemplate}" Header="Description" Binding="{Binding Description, UpdateSourceTrigger=LostFocus}" />
-            </DataGrid.Columns>
-        </DataGrid>
-        
-        <Border Background="#66FFFFFF"
-        Visibility="{Binding IsSelectingNewLocation, Converter={StaticResource BoolToVisibilityConverter}}"
-        Panel.ZIndex="10" Grid.Row="1">
-            <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center">
-                <ProgressBar Width="140" Height="6" IsIndeterminate="True"/>
-                <TextBlock Text="Vyber novou lokaci..." Margin="0,8,0,0" HorizontalAlignment="Center"/>
-            </StackPanel>
+        <Border Grid.Row="0" Padding="0,0,0,12">
+            <ContentControl Content="{Binding LocationDetail}" />
         </Border>
 
-        
+        <Border Grid.Row="1">
+            <ContentControl Content="{Binding ItemsList}" />
+        </Border>
     </Grid>
 </UserControl>
